+++ package/types/repl/util/functional.d.ts
@@ -1,0 +1,2 @@
+export declare function debounce<F extends (...args: any[]) => void | Promise<void>>(func: F, duration?: number): (...args: Parameters<F>) => void | Promise<void>;
+export declare function timeout(callback: () => void, duration?: number): () => void;